fix: paths corrected
