+++ Pokemon/src/pages/main/components/list/full-list.tsx
@@ -1,10 +1,9 @@
 import { useContext, useEffect } from 'react';
+import { useSearchParams } from 'react-router';
 import styles from '../styles/full-list.module.css';
 import { List } from '.';
 import { CardContext } from '../../../../shared';
-import { Pagination } from '../pagination';
-import { Card } from '..';
-import { useSearchParams } from 'react-router';
+import { Pagination, Card } from '..';
 
 export function FullList() {
   const {
