+++ Pokemon/src/shared/index.ts
@@ -6,3 +6,14 @@ export {
   StyleContext,
 } from './context';
 export { changeTheme } from './change-theme';
+export {
+  useGetPokemonQuery,
+  pokemonApi,
+  pokemonReducer,
+  setPokemon,
+  deletePokemon,
+  pokemonSlice,
+  store,
+  useAppDispatch,
+  useAppSelector,
+} from './store';
