+++ Pokemon/src/shared/store/index.ts
@@ -1,19 +1,8 @@
-import { configureStore } from '@reduxjs/toolkit';
-import { useDispatch, useSelector } from 'react-redux';
-import pokemonReducer from './slices/pokemonSlice';
-import { pokemonApi } from './services/api';
-
-export const store = configureStore({
-  reducer: {
-    pokemon: pokemonReducer,
-    [pokemonApi.reducerPath]: pokemonApi.reducer,
-  },
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware().concat(pokemonApi.middleware),
-});
-
-type RootState = ReturnType<typeof store.getState>;
-type AppDispatch = typeof store.dispatch;
-
-export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
-export const useAppSelector = useSelector.withTypes<RootState>();
+export { useGetPokemonQuery, pokemonApi } from './services';
+export {
+  pokemonReducer,
+  setPokemon,
+  deletePokemon,
+  pokemonSlice,
+} from './slices';
+export { store, useAppDispatch, useAppSelector } from './configureStore';
