+++ Pokemon/src/shared/store/services/index.ts
@@ -1,0 +1,1 @@
+export { useGetPokemonQuery, pokemonApi } from './api';
