+++ Pokemon/src/shared/store/slices/index.ts
@@ -1,0 +1,6 @@
+export {
+  pokemonReducer,
+  setPokemon,
+  deletePokemon,
+  pokemonSlice,
+} from './pokemonSlice';
